z-coord
