+++ run.py
@@ -1,22 +1,31 @@
 import os
 from threading import Thread
 import subprocess
+from sgleft import iswinl
+from sgright import iswinr, func2
 
-t1 = Thread(target=subprocess.run, args=(["python", "sg-left.py"],))
-t2 = Thread(target=subprocess.run, args=(["python", "sg-right.py"],))
+#t1 = Thread(target=subprocess.run, args=(["python", "sgleft.py"],))
+t2 = Thread(target=subprocess.run, args=(["python", "sgright.py"],))
 
-t1.start()
+#t1.start()
 t2.start()
-
+while True:
+    print(iswinr())
+    if iswinl():
+        print('LEFT WON!')
+        break
+    if iswinr():
+        print('RIGHT WON!')
+        break
 
 
 
 
 
 '''
-os.system('python sg-left.py && python sg-right.py')
+os.system('python sgleft.py && python sgright.py')
 
-os.system('python sg-right.py')
+os.system('python sgright.py')
 '''
